backend/requirements.txt
aiofile==3.7.4
aiofiles==0.8.0
aiohttp==3.8.1
aiosignal==1.2.0
amqp==5.1.1
anyio==3.5.0
arrow==1.2.2
asgiref==3.5.0
async-timeout==4.0.2
async-unzip==0.2.1
asyncpg==0.26.0
attrs==21.4.0
billiard==3.6.4.0
caio==0.9.5
celery==5.2.6
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
databases==0.6.1
Deprecated==1.2.13
elastic-transport==8.1.1
elasticsearch==8.1.3
et-xmlfile==1.1.0
fastapi==0.75.2
flake8==4.0.1
flower==1.0.0
frozenlist==1.3.0
greenlet==1.1.3
h11==0.13.0
httptools==0.4.0
humanize==4.0.0
idna==3.3
isort==5.10.1
Jinja2==3.1.2
kombu==5.2.4
MarkupSafe==2.1.1
mccabe==0.6.1
multidict==6.0.2
numpy==1.22.3
openpyxl==3.0.9
orjson==3.6.8
packaging==21.3
pandas==1.4.2
prometheus-client==0.14.1
prompt-toolkit==3.0.29
psycopg2-binary==2.9.3
pycodestyle==2.8.0
pydantic==1.9.0
pyflakes==2.4.0
pyparsing==3.0.8
python-dateutil==2.8.2
python-dotenv==0.20.0
python-multipart==0.0.5
pytz==2022.1
PyYAML==6.0
redis==4.2.2
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.40
starlette==0.17.1
tornado==6.1
typing_extensions==4.2.0
urllib3==1.26.9
uvicorn==0.17.6
uvloop==0.16.0
vine==5.0.0
watchgod==0.8.2
wcwidth==0.2.5
websockets==10.3
wrapt==1.14.0
xlrd==2.0.1
XlsxWriter==3.0.3
yarl==1.7.2
